old_pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.44",
    "tabulate>=0.9.0",
]
